requirements.txt
discord.py
jishaku
python-dotenv
pyicu